--- app/src/disclosure/index.ts
@@ -1,1 +1,0 @@
-export { Disclosure, type DisclosureProps } from './Disclosure';
