+++ app/src/pages/DisclosurePage.tsx
@@ -1,16 +1,15 @@
 import { Disclosure } from '@moondreamsdev/dreamer-ui/components';
-import { ComponentPage } from '../../components/layout/ComponentPage';
-
-const tableOfContents = [
-  { id: 'showcase', title: 'Component Showcase', level: 1 },
-  { id: 'default', title: 'Default Variant', level: 2 },
-  { id: 'danger', title: 'Danger Area', level: 2 },
-  { id: 'styled', title: 'Styled Button', level: 2 },
-  { id: 'disabled', title: 'Disabled State', level: 2 },
-  { id: 'toc', title: 'Table of Contents', level: 2 },
-];
+import { ComponentPage } from '../components/layout/ComponentPage';
 
 export const DisclosurePage = () => {
+  const tableOfContents = [
+    { id: 'usage', title: 'Usage', level: 2 },
+    { id: 'showcase', title: 'Component Showcase', level: 2 },
+    { id: 'default', title: 'Default Variant', level: 3 },
+    { id: 'danger', title: 'Danger Area', level: 3 },
+    { id: 'styled', title: 'Styled Button', level: 3 },
+    { id: 'disabled', title: 'Disabled State', level: 3 },
+  ];
   return (
     <ComponentPage
       title='Disclosure'
@@ -19,7 +18,7 @@ export const DisclosurePage = () => {
     >
       <div className='space-y-8'>
         {/* Usage Notice */}
-        <div className='bg-blue-900/20 border border-blue-700 rounded-lg p-4'>
+        <div id='usage' className='bg-blue-900/20 border border-blue-700 rounded-lg p-4'>
           <div className='flex items-center space-x-2'>
             <svg className='w-5 h-5 text-blue-400' fill='currentColor' viewBox='0 0 20 20'>
               <path
@@ -80,29 +79,7 @@ export const DisclosurePage = () => {
             <div id='disabled' className='mb-6'>
               <h3 className='text-lg font-medium text-white mb-3'>Disabled State</h3>
               <Disclosure label='Disabled disclosure' disabled>
-                <div className='p-2 text-gray-300 hover:text-white hover:bg-gray-700/50'>
-                  This disclosure is disabled and cannot be opened.
-                </div>
-              </Disclosure>
-            </div>
-
-            {/* Table of contents */}
-            <div id='toc' className='mb-6'>
-              <h3 className='text-lg font-medium text-white mb-3'>Table of Contents</h3>
-              <Disclosure
-                label='Table of Contents'
-                className='bg-gray-800/50 border border-accent-medium/50 rounded-xl text-white w-fit mx-auto'
-                buttonClassName='p-4 text-left'
-              >
-                <nav className='p-2 pt-1 flex flex-col *:text-left *:p-1 *:hover:font-medium'>
-                  <button className='text-gray-300 hover:text-white'>Introduction</button>
-                  <button className='text-gray-300 hover:text-white'>Section 1</button>
-                  <button className='text-gray-300/60 hover:text-white/60 ml-3'>Subsection A</button>
-                  <button className='text-gray-300/60 hover:text-white/60 ml-3'>Subsection B</button>
-                  <button className='text-gray-300 hover:text-white'>Section 2</button>
-                  <button className='text-gray-300 hover:text-white'>Section 3</button>
-                  <button className='text-gray-300 hover:text-white'>Section 4</button>
-                </nav>
+                <div className='p-2 text-gray-400'>This disclosure is disabled and cannot be opened.</div>
               </Disclosure>
             </div>
           </div>
