+++ lib/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@moondreamsdev/dreamer-ui",
-  "version": "1.7.10-test.6",
+  "version": "1.7.10-test.8",
   "publishConfig": {
     "access": "public"
   },
