+++ lib/src/components/disclosure/index.ts
@@ -1,0 +1,1 @@
+export { Disclosure, type DisclosureProps } from './Disclosure';
