+++ lib/src/symbols/CaretExpand.tsx
@@ -1,4 +1,4 @@
-import { SymbolProps } from '@moondreamsdev/dreamer-ui/symbols';
+import { SymbolProps } from './props';
 
 export function CaretExpand({ size = 15, color = 'currentColor', className = 'inline', ...props }: SymbolProps) {
   return (
@@ -12,27 +12,6 @@ export function CaretExpand({ size = 15, color = 'currentColor', className = 'in
     >
       <path
         d='M4.93179 5.43179C4.75605 5.60753 4.75605 5.89245 4.93179 6.06819C5.10753 6.24392 5.39245 6.24392 5.56819 6.06819L7.49999 4.13638L9.43179 6.06819C9.60753 6.24392 9.89245 6.24392 10.0682 6.06819C10.2439 5.89245 10.2439 5.60753 10.0682 5.43179L7.81819 3.18179C7.73379 3.0974 7.61933 3.04999 7.49999 3.04999C7.38064 3.04999 7.26618 3.0974 7.18179 3.18179L4.93179 5.43179ZM10.0682 9.56819C10.2439 9.39245 10.2439 9.10753 10.0682 8.93179C9.89245 8.75606 9.60753 8.75606 9.43179 8.93179L7.49999 10.8636L5.56819 8.93179C5.39245 8.75606 5.10753 8.75606 4.93179 8.93179C4.75605 9.10753 4.75605 9.39245 4.93179 9.56819L7.18179 11.8182C7.35753 11.9939 7.64245 11.9939 7.81819 11.8182L10.0682 9.56819Z'
-        fill={color}
-        fillRule='evenodd'
-        clipRule='evenodd'
-      />
-    </svg>
-  );
-}
-
-export function CaretCollapse({ size = 15, color = 'currentColor', className = 'inline', ...props }: SymbolProps) {
-  return (
-    <svg
-      {...props}
-      width={size}
-      height={size}
-      className={className}
-      viewBox='0 0 15 15'
-      xmlns='http://www.w3.org/2000/svg'
-    >
-      <path d='M4.5 4 L7.5 7 L10.5 4 L9.8 3.3 L7.5 5.6 L5.2 3.3 Z' fill={color} fillRule='evenodd' clipRule='evenodd' />
-      <path
-        d='M4.5 11 L7.5 8 L10.5 11 L9.8 11.7 L7.5 9.4 L5.2 11.7 Z'
         fill={color}
         fillRule='evenodd'
         clipRule='evenodd'
